Remove logs from build, clean up `const.ts` and the env file loading process, consolidate env variables
