+++ src/typescript/frontend/.env.example
@@ -4,3 +4,9 @@ NEXT_PUBLIC_APTOS_NETWORK="local"
 # Whether or not revalidation times should be short. This should be set to "true" for local development
 # and "false" for production.
 NEXT_PUBLIC_SHORT_REVALIDATE="true"
+
+# Whether or not to fetch static JSON files, likely in a staging environment.
+NEXT_PUBLIC_FORCE_STATIC_FETCH="true"
+
+# The address of the main `emojicoin_dot_fun` contract module.
+NEXT_PUBLIC_MODULE_ADDRESS="4bab58978ec1b1bef032eeb285ad47a6a9b997d646c19b598c35f46b26ff9ece"
