+++ src/typescript/frontend/.env.production
@@ -1,3 +1,4 @@
 NEXT_PUBLIC_APTOS_NETWORK="local"
 NEXT_PUBLIC_SHORT_REVALIDATE="false"
 NEXT_PUBLIC_FORCE_STATIC_FETCH="true"
+NEXT_PUBLIC_MODULE_ADDRESS="4bab58978ec1b1bef032eeb285ad47a6a9b997d646c19b598c35f46b26ff9ece"
