+++ src/typescript/frontend/package.json
@@ -63,7 +63,7 @@
     "build": "next build",
     "build-dev": "chmod +x ./scripts/build-dev.sh && ./scripts/build-dev.sh",
     "build-prod": "chmod +x ./scripts/build-prod.sh && ./scripts/build-prod.sh",
-    "build:debug": "export BUILD_DEBUG=true && pnpm clean && next build --no-lint --no-mangling --debug && export BUILD_DEBUG=\"\"",
+    "build:debug": "BUILD_DEBUG=true && pnpm clean && next build --no-lint --no-mangling --debug",
     "clean": "rm -rf .next",
     "dev": "NODE_OPTIONS='--inspect' next dev --turbo",
     "format": "pnpm _format --write",
