+++ src/typescript/frontend/src/app/emojicoin/[market]/page.tsx
@@ -1,7 +1,6 @@
 import Emojicoin from "components/pages/emojicoin";
 import { SHORT_REVALIDATE } from "lib/env";
 import getInitialChatData from "lib/queries/initial/chats";
-import { SAMPLE_DATA_BASE_URL } from "lib/queries/initial/const";
 import { fetchTopMarkets } from "lib/queries/initial/markets";
 import { fetchLastMarketState } from "lib/queries/initial/state";
 import getInitialSwapData from "lib/queries/initial/swaps";
@@ -14,13 +13,6 @@ type StaticParams = {
 };
 
 export const generateStaticParams = async (): Promise<Array<StaticParams>> => {
-  const res = await fetch(new URL("top-market-data.json", SAMPLE_DATA_BASE_URL));
-  console.warn("-".repeat(80));
-  console.warn(res);
-  const data2 = await res.json();
-  console.warn(new URL("top-market-data.json", SAMPLE_DATA_BASE_URL));
-  console.warn(res);
-  console.warn(data2);
   const data = await fetchTopMarkets();
   return data.map(v => ({
     market: v.state.marketMetadata.marketID.toString(),
@@ -35,17 +27,6 @@ interface EmojicoinPageProps {
 const EmojicoinPage = async (params: EmojicoinPageProps) => {
   const marketID = params.params.market;
   const res = await fetchLastMarketState(marketID);
-
-  console.warn("Building page", res, marketID, params, params.params.market, params.searchParams, params.params, params.searchParams);
-
-{  const url = new URL(`chat-data-${Number(marketID)}.json`, SAMPLE_DATA_BASE_URL);
-  const r = await fetch(url);
-  const d = await r.json();
-  console.warn(url, r, d);}
-{  const url = new URL(`swap-data-${Number(marketID)}.json`, SAMPLE_DATA_BASE_URL);
-  const r = await fetch(url);
-  const d = await r.json();
-  console.warn(url, r, d);}
 
   if (res) {
     const chatData = await getInitialChatData(marketID);
