+++ src/typescript/sdk/tests/utils/index.ts
@@ -5,3 +5,4 @@ export * from "./helpers";
 export * from "./local-node";
 export * from "./publish";
 export * from "./types";
+export * from "./load-priv-key";
